orfd/requirements.txt
flask
bert-serving-server
bert-serving-client
requests
BeautifulSoup4
pandas
numpy
sklearn
apscheduler